Declutter queryParams, introduce 'mode'
Having different parameters in the URL which actually are just values
for the same attribute is bad practice. In reality all of 'test',
'client-app', etc. were just indicating the mode in which the app is
started and as such should be values of a fixed parameter.

The app should be started with index.html?mode=<value>.